requirements.txt
streamlit==1.32.0
google-generativeai==0.5.4
gspread==6.1.2
google-auth==2.29.0
pdfplumber==0.11.0
python-dotenv==1.0.1
plotly==5.20.0
pandas==2.2.2
